requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
pyodbc==5.0.1
openai==1.3.0
Werkzeug==2.3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.6
blinker==1.6.2
waitress==2.1.2
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.1